Rename plugin to better match its use
