+++ plugin/exceptions.py
@@ -1,0 +1,2 @@
+class UnknownClassException(Exception):
+    pass
