+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "mkdocs-opensafely-backend-contracts"
+name = "mkdocs-opensafely-databuilder"
 description = ""
 version = "1+local"
 readme = "README.md"
@@ -16,12 +16,12 @@ dependencies = [
 ]
 
 [project.entry-points."mkdocs.plugins"]
-contracts = "contracts:BackendContractsPlugin"
+"mkdocs-opensafely-databuilder" = "plugin.main:DataBuilderPlugin"
 
 [project.urls]
 Home = "https://opensafely.org"
 Documentation = "https://docs.opensafely.org"
-Source = "https://github.com/opensafely-core/mkdocs-opensafely-backend-contracts"
+Source = "https://github.com/opensafely-core/mkdocs-opensafely-databuilder"
 
 [tool.black]
 exclude = '''
@@ -50,7 +50,7 @@ lines_after_imports = 2
 skip_glob = [".direnv", "venv", ".venv"]
 
 [tool.flit.module]
-name = "contracts"
+name = "plugin"
 
 [tool.pytest.ini_options]
 addopts = "--tb=native --strict-markers"
