+++ requirements.prod.txt
@@ -7,4 +7,4 @@
 first==2.0.2 \
     --hash=sha256:8d8e46e115ea8ac652c76123c0865e3ff18372aef6f03c22809ceefcea9dec86 \
     --hash=sha256:ff285b08c55f8c97ce4ea7012743af2495c9f1291785f163722bd36f6af6d3bf
-    # via mkdocs-opensafely-backend-contracts (pyproject.toml)
+    # via mkdocs-opensafely-databuilder (pyproject.toml)
